Fixed a glaringly awful syntax error. Merged EFI code back in.

+++ include/sdk/feral/feralobjs.h
@@ -6,12 +6,12 @@
 #include <stddef.h>
 
 #if defined(__x86_64__) || defined(__i386__)
-typedef unsigned long  UINT64;
+typedef unsigned long long  UINT64;
 typedef unsigned int   UINT32;
 typedef unsigned short UINT16;
 typedef unsigned char  UINT8;
 	
-typedef signed long  INT64;
+typedef signed long long  INT64;
 typedef signed int   INT32;
 typedef signed short INT16;	
 typedef signed char  INT8;
@@ -21,12 +21,12 @@ typedef UINT8 BYTE;
 #endif
 
 #if defined(__arm__) || defined(__aarch64__)
-typedef unsigned long  UINT64;
+typedef unsigned long long  UINT64;
 typedef unsigned int   UINT32;
 typedef unsigned short UINT16;
 typedef unsigned char  UINT8;
 	
-typedef signed long  INT64;
+typedef signed long long  INT64;
 typedef signed int   INT32;
 typedef signed short INT16;	
 typedef signed char  INT8;
